Refactor: move tests into same directory and remove packages
